feat containers/vector: move vector to containers/dynamic
