+++ src/containers/CMakeLists.txt
@@ -1,7 +1,6 @@
 target_sources(${PROJECT_NAME}
     PRIVATE
         common.h
-        vector.h
         array.h
 )
