+++ src/containers/dynamic/CMakeLists.txt
@@ -4,3 +4,4 @@ target_sources(${PROJECT_NAME}
 )
 
 add_subdirectory(string)
+add_subdirectory(vector)
